services/fashion-clip/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.40.0
torch==2.3.0
Pillow==10.3.0
requests==2.31.0
pydantic==2.7.0
pytest==7.4.0
httpx==0.27.0
rembg==2.0.57
onnxruntime==1.17.0
